requirements.txt
openai
google-generativeai
streamlit
graphviz
